Enable logging again + bugfix for an issue a customer had.
The first time a service object is executed and the service instance is in impersonation mode, the ServiceobjectToType property is going to be initialized by callin gthe DescribeServiceObjects() on the various serviceObjectClasses. Because we're in impersonation mode AND we're therefor executing in the user context, the DescribeServiceObject of the ProcessInstanceClientSO threw an exception which cause sthe rest of the objects to fail and load.